Fixes bug with following/unfollowing system theme
